+++ packages/agtree/test/parser/misc/modifier-list.test.ts
@@ -4,8 +4,6 @@ import { ModifierListParser } from '../../../src/parser/misc/modifier-list';
 import { type ModifierList } from '../../../src/nodes';
 import { EMPTY, SPACE } from '../../../src/utils/constants';
 import { ModifierListGenerator } from '../../../src/generator/misc/modifier-list-generator';
-import { ModifierListSerializer } from '../../../src/serializer/misc/modifier-list-serializer';
-import { ModifierListDeserializer } from '../../../src/deserializer/misc/modifier-list-deserializer';
 
 describe('ModifierListParser', () => {
     test('parse', () => {
@@ -751,37 +749,5 @@ describe('ModifierListParser', () => {
                 '~path=/\\/(sub1|sub2)\\/page\\.html/,replace=/(<VAST[\\s\\S]*?>)[\\s\\S]*<\\/VAST>/\\$1<\\/VAST>/i',
             ),
         ).toEqual('~path=/\\/(sub1|sub2)\\/page\\.html/,replace=/(<VAST[\\s\\S]*?>)[\\s\\S]*<\\/VAST>/\\$1<\\/VAST>/i');
-    });
-
-    describe('serialize & deserialize', () => {
-        test.each([
-            // single modifier
-            'foo',
-            '~foo',
-            'foo=bar',
-            '~foo=bar',
-
-            // multiple modifiers
-            'foo,bar',
-            'foo,~bar',
-            '~foo,bar',
-            '~foo,~bar',
-
-            'foo=bar,bar=foo',
-            '~foo=bar,~bar=foo',
-            'foo=bar,~bar=foo',
-            '~foo=bar,bar=foo',
-
-            // complicated
-            'path=/\\/(sub1|sub2)\\/page\\.html/,replace=/(<VAST[\\s\\S]*?>)[\\s\\S]*<\\/VAST>/\\$1<\\/VAST>/i',
-            'foo=你好,bar=世界',
-        ])("should serialize and deserialize '%p'", async (input) => {
-            await expect(input).toBeSerializedAndDeserializedProperly(
-                ModifierListParser,
-                ModifierListGenerator,
-                ModifierListSerializer,
-                ModifierListDeserializer,
-            );
-        });
     });
 });